Moved types in folder Converter into matching namespace
