+++ FileHelpers.Examples/Examples/18.Converters/60.CustomConverter.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using FileHelpers;
+using FileHelpers.Converters;
 
 namespace ExamplesFx
 {
