+++ FileHelpers.Tests/Tests/Common/CustomConverter3.cs
@@ -1,3 +1,4 @@
+using FileHelpers.Converters;
 using NUnit.Framework;
 
 namespace FileHelpers.Tests.CommonTests
