+++ FileHelpers.Tests/Tests/Errors/BadUsage.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using FileHelpers.Converters;
 using NUnit.Framework;
 
 namespace FileHelpers.Tests.Errors
